requirements.txt
# Streamlit for web UI
streamlit>=1.28.0

# Vector database
chromadb>=0.4.15

# Embeddings - Fixed compatible versions
sentence-transformers>=2.2.2,<3.0.0
transformers>=4.21.0,<5.0.0
torch>=1.9.0
huggingface-hub>=0.16.0,<0.20.0
safetensors>=0.3.1
openai>=1.3.0

# Configuration and environment
PyYAML>=6.0
python-dotenv>=1.0.0

# Text processing
tiktoken>=0.5.1

# File monitoring
watchdog>=3.0.0

# Git integration
GitPython>=3.1.37

# Code parsing with tree-sitter
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0
tree-sitter-javascript>=0.21.0
tree-sitter-typescript>=0.21.0
tree-sitter-java>=0.21.0
tree-sitter-cpp>=0.21.0
tree-sitter-c>=0.21.0
tree-sitter-go>=0.21.0
tree-sitter-rust>=0.21.0

# LangChain for text processing
langchain>=0.0.300

# Anthropic LLM support
anthropic>=0.7.0 